backend/requirements_error_handling.txt
# Additional requirements for enhanced error handling
# Add these to your existing requirements.txt

# JWT handling for enhanced authentication
PyJWT==2.8.0

# Enhanced logging (optional but recommended)
structlog==23.2.0

# For better error tracking and monitoring (optional)
sentry-sdk[flask]==1.38.0

# For rate limiting with Redis (optional but recommended for production)
redis==5.0.1
flask-redis==0.4.0

# For request validation (optional enhancement)
marshmallow==3.20.2
flask-marshmallow==0.15.0

# For API documentation (optional but helpful)
flask-restx==1.3.0

# For better CORS handling
flask-cors==4.0.0

# For environment variable management
python-dotenv==1.0.0

# Database connection pooling (recommended for production)
psycopg2-pool==1.1
